# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CustomerInput {
  nome: string;
  /** @nullable */
  fazenda?: string | null;
  /** @nullable */
  cpfCnpj?: string | null;
  /** @nullable */
  inscricao?: string | null;
  /** @nullable */
  endereco?: string | null;
  /** @nullable */
  uf?: string | null;
  /** @nullable */
  municipio?: string | null;
  /** @nullable */
  telefone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  roteiro?: string | null;
  /** @nullable */
  banco?: string | null;
  /** @nullable */
  conta?: string | null;
  /** @nullable */
  agencia?: string | null;
  /** @nullable */
  titular?: string | null;
}

export type Customer = CustomerInput & {
  id: number;
  createdAt: string;
};

export interface OrderItemInput {
  especie: string;
  raca: string;
  idade: string;
  quantidade: number;
  peso: number;
  precoArroba: number;
  rastreabilidade: boolean;
}

export type OrderItem = OrderItemInput & {
  id: number;
  total: number;
};

export interface OrderInput {
  compradorId: number;
  vendedorId: number;
  /** @nullable */
  dataEmbarque?: string | null;
  /** @nullable */
  dataDesembarque?: string | null;
  /** @nullable */
  dataAbate?: string | null;
  /** @nullable */
  roteiro?: string | null;
  formaPagamento: string;
  /** @nullable */
  prazoPagamento?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  dataCompra?: string | null;
  /** @nullable */
  status?: string | null;
  items: OrderItemInput[];
}

export interface Order {
  id: number;
  numero: string;
  dataEmissao: string;
  compradorId: number;
  vendedorId: number;
  comprador?: Customer;
  vendedor?: Customer;
  /** @nullable */
  dataEmbarque?: string | null;
  /** @nullable */
  dataDesembarque?: string | null;
  /** @nullable */
  dataAbate?: string | null;
  /** @nullable */
  roteiro?: string | null;
  formaPagamento: string;
  /** @nullable */
  prazoPagamento?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  dataCompra?: string | null;
  status: string;
  totalAnimais: number;
  totalPeso: number;
  totalValor: number;
  items: OrderItem[];
}

export interface OrderSummary {
  id: number;
  numero: string;
  dataEmissao: string;
  compradorNome: string;
  vendedorNome: string;
  /** @nullable */
  municipio?: string | null;
  totalAnimais: number;
  totalValor: number;
  status: string;
}

export interface DashboardSummary {
  totalPedidos: number;
  totalAnimais: number;
  totalValor: number;
  totalCompradores: number;
  totalVendedores: number;
  pedidosMes: number;
  valorMes: number;
}

export interface GroupBucket {
  label: string;
  animais: number;
  valor: number;
}

export interface MonthlyBucket {
  month: string;
  pedidos: number;
  animais: number;
  valor: number;
}

export interface RankingEntry {
  nome: string;
  pedidos: number;
  animais: number;
  valor: number;
}

export type ListOrdersParams = {
  status?: string;
  customerId?: number;
  from?: string;
  to?: string;
};