# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerInput } from "./customerInput";

export type Customer = CustomerInput & {
  id: number;
  createdAt: Date;
};