# lib/api-zod/src/generated/types/customerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerInput {
  nome: string;
  /** @nullable */
  fazenda?: string | null;
  /** @nullable */
  cpfCnpj?: string | null;
  /** @nullable */
  inscricao?: string | null;
  /** @nullable */
  endereco?: string | null;
  /** @nullable */
  uf?: string | null;
  /** @nullable */
  municipio?: string | null;
  /** @nullable */
  telefone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  roteiro?: string | null;
  /** @nullable */
  banco?: string | null;
  /** @nullable */
  conta?: string | null;
  /** @nullable */
  agencia?: string | null;
  /** @nullable */
  titular?: string | null;
}