# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalPedidos: number;
  totalAnimais: number;
  totalValor: number;
  totalCompradores: number;
  totalVendedores: number;
  pedidosMes: number;
  valorMes: number;
}