# lib/api-zod/src/generated/types/groupBucket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupBucket {
  label: string;
  animais: number;
  valor: number;
}