# lib/api-zod/src/generated/types/monthlyBucket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyBucket {
  month: string;
  pedidos: number;
  animais: number;
  valor: number;
}