# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  numero: string;
  dataEmissao: Date;
  compradorId: number;
  vendedorId: number;
  comprador?: Customer;
  vendedor?: Customer;
  /** @nullable */
  dataEmbarque?: string | null;
  /** @nullable */
  dataDesembarque?: string | null;
  /** @nullable */
  dataAbate?: string | null;
  /** @nullable */
  roteiro?: string | null;
  formaPagamento: string;
  /** @nullable */
  prazoPagamento?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  dataCompra?: string | null;
  status: string;
  totalAnimais: number;
  totalPeso: number;
  totalValor: number;
  items: OrderItem[];
}