# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from "./orderItemInput";

export interface OrderInput {
  compradorId: number;
  vendedorId: number;
  /** @nullable */
  dataEmbarque?: string | null;
  /** @nullable */
  dataDesembarque?: string | null;
  /** @nullable */
  dataAbate?: string | null;
  /** @nullable */
  roteiro?: string | null;
  formaPagamento: string;
  /** @nullable */
  prazoPagamento?: string | null;
  /** @nullable */
  observacao?: string | null;
  /** @nullable */
  dataCompra?: string | null;
  /** @nullable */
  status?: string | null;
  items: OrderItemInput[];
}