# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  especie: string;
  raca: string;
  idade: string;
  quantidade: number;
  peso: number;
  precoArroba: number;
  rastreabilidade: boolean;
}