# lib/api-zod/src/generated/types/orderSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EP Bov Pedidos API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderSummary {
  id: number;
  numero: string;
  dataEmissao: Date;
  compradorNome: string;
  vendedorNome: string;
  /** @nullable */
  municipio?: string | null;
  totalAnimais: number;
  totalValor: number;
  status: string;
}